requirements.txt
kivy==2.3.0
kivymd==1.1.1
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
pytesseract==0.3.10
pyjnius==1.5.0
plyer==2.1.0
cython==0.29.33